Rename fah_plotting -> plots
